# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.ResponseWithNoBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.ResponseWithStringData:
    properties:
      code:
        type: string
      data:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: Swagger API for Golang Project PMS + Freelancer.
  license: {}
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /uploadAttachment/{id}:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Upload File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Uploading Attachment here
      tags:
      - Project
  /uploadPicture:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Upload File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithStringData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Uploading Picture here
      tags:
      - User
...
swagger: "2.0"